refactor(agent-core): extract lib test sections into modules (#1259)
