+++ crates/tau-agent-core/src/lib.rs
@@ -3029,3120 +3029,15 @@ mod tests {
         }
     }
 
-    #[test]
-    fn unit_agent_config_defaults_include_request_and_tool_timeouts() {
-        let config = AgentConfig::default();
-        assert_eq!(config.agent_id, "tau-agent");
-        assert_eq!(config.request_timeout_ms, Some(120_000));
-        assert_eq!(config.tool_timeout_ms, Some(120_000));
-        assert!(config.stream_retry_with_buffering);
-        assert_eq!(config.max_estimated_input_tokens, Some(120_000));
-        assert_eq!(config.max_estimated_total_tokens, None);
-        assert_eq!(config.structured_output_max_retries, 1);
-        assert_eq!(config.react_max_replans_on_tool_failure, 1);
-        assert_eq!(config.memory_retrieval_limit, 3);
-        assert_eq!(config.memory_embedding_dimensions, 128);
-        assert_eq!(config.memory_min_similarity, 0.55);
-        assert_eq!(config.memory_max_chars_per_item, 180);
-        assert_eq!(config.memory_embedding_model, None);
-        assert_eq!(config.memory_embedding_api_base, None);
-        assert_eq!(config.memory_embedding_api_key, None);
-        assert!(config.response_cache_enabled);
-        assert_eq!(config.response_cache_max_entries, 128);
-        assert!(config.tool_result_cache_enabled);
-        assert_eq!(config.tool_result_cache_max_entries, 256);
-        assert_eq!(config.model_input_cost_per_million, None);
-        assert_eq!(config.model_output_cost_per_million, None);
-        assert_eq!(config.cost_budget_usd, None);
-        assert_eq!(config.cost_alert_thresholds_percent, vec![80, 100]);
-        assert_eq!(config.async_event_queue_capacity, 128);
-        assert_eq!(config.async_event_handler_timeout_ms, Some(5_000));
-        assert!(!config.async_event_block_on_full);
-    }
+    #[path = "config_and_direct_message.rs"]
+    mod config_and_direct_message;
 
-    #[test]
-    fn unit_cache_insert_with_limit_evicts_oldest_entries() {
-        let mut cache = HashMap::new();
-        let mut order = VecDeque::new();
-        cache_insert_with_limit(&mut cache, &mut order, "a".to_string(), 1_u32, 2);
-        cache_insert_with_limit(&mut cache, &mut order, "b".to_string(), 2_u32, 2);
-        cache_insert_with_limit(&mut cache, &mut order, "c".to_string(), 3_u32, 2);
+    #[path = "async_event_dispatch.rs"]
+    mod async_event_dispatch;
 
-        assert_eq!(cache.len(), 2);
-        assert!(!cache.contains_key("a"));
-        assert_eq!(cache.get("b"), Some(&2));
-        assert_eq!(cache.get("c"), Some(&3));
-    }
+    #[path = "streaming_and_budgets.rs"]
+    mod streaming_and_budgets;
 
-    #[test]
-    fn unit_stream_retry_buffer_on_delta_suppresses_replayed_prefix() {
-        let mut state = StreamingRetryBufferState::default();
-        assert_eq!(
-            stream_retry_buffer_on_delta(&mut state, "Hel"),
-            Some("Hel".to_string())
-        );
-
-        state.reset_attempt();
-        assert_eq!(stream_retry_buffer_on_delta(&mut state, "Hel"), None);
-        assert_eq!(
-            stream_retry_buffer_on_delta(&mut state, "lo"),
-            Some("lo".to_string())
-        );
-    }
-
-    #[test]
-    fn unit_dynamic_tool_registry_supports_presence_and_lifecycle_helpers() {
-        let mut agent = Agent::new(Arc::new(EchoClient), AgentConfig::default());
-        assert!(!agent.has_tool("read"));
-        assert!(!agent.unregister_tool("read"));
-
-        agent.register_tool(ReadTool);
-        assert!(agent.has_tool("read"));
-        assert_eq!(agent.registered_tool_names(), vec!["read".to_string()]);
-
-        assert!(agent.unregister_tool("read"));
-        assert!(!agent.has_tool("read"));
-
-        agent.register_tool(ReadTool);
-        agent.clear_tools();
-        assert!(agent.registered_tool_names().is_empty());
-    }
-
-    #[test]
-    fn unit_direct_message_policy_enforces_configured_routes() {
-        let mut policy = AgentDirectMessagePolicy::default();
-        assert!(!policy.allows("planner", "executor"));
-        assert!(!policy.allows("planner", "planner"));
-
-        policy.allow_route("planner", "executor");
-        assert!(policy.allows("planner", "executor"));
-        assert!(!policy.allows("executor", "planner"));
-
-        policy.allow_bidirectional_route("reviewer", "executor");
-        assert!(policy.allows("reviewer", "executor"));
-        assert!(policy.allows("executor", "reviewer"));
-
-        policy.allow_self_messages = true;
-        assert!(policy.allows("planner", "planner"));
-    }
-
-    #[test]
-    fn functional_send_direct_message_appends_system_message() {
-        let sender = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "planner".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let mut recipient = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "executor".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let mut policy = AgentDirectMessagePolicy::default();
-        policy.allow_route("planner", "executor");
-
-        sender
-            .send_direct_message(&mut recipient, "  review this step  ", &policy)
-            .expect("direct message should be accepted");
-
-        let direct_message = recipient
-            .messages()
-            .iter()
-            .find(|message| {
-                message.role == MessageRole::System
-                    && message.text_content().starts_with(DIRECT_MESSAGE_PREFIX)
-            })
-            .expect("direct message should be appended as a system message");
-        assert!(direct_message
-            .text_content()
-            .contains("from=planner to=executor"));
-        assert!(direct_message.text_content().contains("review this step"));
-    }
-
-    #[tokio::test]
-    async fn integration_direct_message_is_included_in_recipient_prompt_context() {
-        let sender = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "planner".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let recipient_client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ack"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut recipient = Agent::new(
-            recipient_client.clone(),
-            AgentConfig {
-                agent_id: "executor".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let mut policy = AgentDirectMessagePolicy::default();
-        policy.allow_route("planner", "executor");
-
-        sender
-            .send_direct_message(&mut recipient, "Focus on retry semantics", &policy)
-            .expect("route should be authorized");
-        let _ = recipient
-            .prompt("continue")
-            .await
-            .expect("recipient prompt should succeed");
-
-        let requests = recipient_client.requests.lock().await;
-        let request = requests.first().expect("captured request");
-        assert!(
-            request.messages.iter().any(|message| {
-                message.role == MessageRole::System
-                    && message
-                        .text_content()
-                        .contains("[Tau direct message] from=planner to=executor")
-            }),
-            "direct message should be included in prompt context"
-        );
-    }
-
-    #[test]
-    fn regression_unauthorized_direct_message_fails_closed_without_mutation() {
-        let sender = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "planner".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let mut recipient = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "executor".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let policy = AgentDirectMessagePolicy::default();
-        let baseline_count = recipient.messages().len();
-
-        let error = sender
-            .send_direct_message(&mut recipient, "unauthorized", &policy)
-            .expect_err("unauthorized route must fail closed");
-        assert!(matches!(
-            error,
-            AgentDirectMessageError::UnauthorizedRoute { .. }
-        ));
-        assert_eq!(recipient.messages().len(), baseline_count);
-    }
-
-    #[test]
-    fn regression_direct_message_policy_enforces_max_message_chars() {
-        let sender = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "planner".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let mut recipient = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                agent_id: "executor".to_string(),
-                ..AgentConfig::default()
-            },
-        );
-        let mut policy = AgentDirectMessagePolicy::default();
-        policy.allow_route("planner", "executor");
-        policy.max_message_chars = 5;
-        let baseline_count = recipient.messages().len();
-
-        let error = sender
-            .send_direct_message(&mut recipient, "message too long", &policy)
-            .expect_err("oversized direct message must fail");
-        assert!(matches!(
-            error,
-            AgentDirectMessageError::MessageTooLong { .. }
-        ));
-        assert_eq!(recipient.messages().len(), baseline_count);
-    }
-
-    #[tokio::test]
-    async fn functional_with_scoped_tool_registers_within_scope_and_restores_after() {
-        let mut agent = Agent::new(Arc::new(EchoClient), AgentConfig::default());
-        assert!(!agent.has_tool("read"));
-
-        let value = agent
-            .with_scoped_tool(ReadTool, |agent| {
-                Box::pin(async move {
-                    assert!(agent.has_tool("read"));
-                    assert_eq!(agent.registered_tool_names(), vec!["read".to_string()]);
-                    42usize
-                })
-            })
-            .await;
-
-        assert_eq!(value, 42);
-        assert!(!agent.has_tool("read"));
-    }
-
-    #[tokio::test]
-    async fn unit_cooperative_cancellation_token_signals_waiters() {
-        let token = CooperativeCancellationToken::new();
-        let waiter = token.clone();
-        let task = tokio::spawn(async move {
-            waiter.cancelled().await;
-            1usize
-        });
-
-        tokio::time::sleep(Duration::from_millis(5)).await;
-        token.cancel();
-
-        assert!(token.is_cancelled());
-        assert_eq!(task.await.expect("waiter task should complete"), 1);
-    }
-
-    #[test]
-    fn unit_estimate_chat_request_tokens_accounts_for_tools_and_max_tokens() {
-        let request = ChatRequest {
-            model: "openai/gpt-4o-mini".to_string(),
-            messages: vec![
-                Message::system("sys"),
-                Message::user("hello world"),
-                Message::assistant_blocks(vec![ContentBlock::ToolCall {
-                    id: "call-1".to_string(),
-                    name: "read".to_string(),
-                    arguments: serde_json::json!({ "path": "README.md" }),
-                }]),
-            ],
-            tools: vec![ToolDefinition {
-                name: "read".to_string(),
-                description: "Read file contents".to_string(),
-                parameters: serde_json::json!({
-                    "type": "object",
-                    "properties": {
-                        "path": { "type": "string" }
-                    }
-                }),
-            }],
-            tool_choice: Some(ToolChoice::Auto),
-            json_mode: false,
-            max_tokens: Some(64),
-            temperature: Some(0.0),
-        };
-
-        let estimate = estimate_chat_request_tokens(&request);
-        assert!(estimate.input_tokens > 0);
-        assert_eq!(
-            estimate.total_tokens,
-            estimate.input_tokens.saturating_add(64)
-        );
-    }
-
-    #[test]
-    fn functional_estimate_chat_request_tokens_accounts_for_media_blocks() {
-        let baseline = ChatRequest {
-            model: "openai/gpt-4o-mini".to_string(),
-            messages: vec![Message::user("hello")],
-            tools: vec![],
-            tool_choice: None,
-            json_mode: false,
-            max_tokens: Some(32),
-            temperature: None,
-        };
-        let with_media = ChatRequest {
-            model: "openai/gpt-4o-mini".to_string(),
-            messages: vec![Message {
-                role: MessageRole::User,
-                content: vec![
-                    ContentBlock::text("hello"),
-                    ContentBlock::image_base64("image/png", "aW1hZ2VEYXRh"),
-                    ContentBlock::audio_base64("audio/wav", "YXVkaW9EYXRh"),
-                ],
-                tool_call_id: None,
-                tool_name: None,
-                is_error: false,
-            }],
-            tools: vec![],
-            tool_choice: None,
-            json_mode: false,
-            max_tokens: Some(32),
-            temperature: None,
-        };
-
-        let baseline_estimate = estimate_chat_request_tokens(&baseline);
-        let media_estimate = estimate_chat_request_tokens(&with_media);
-        assert!(media_estimate.input_tokens > baseline_estimate.input_tokens);
-        assert!(media_estimate.total_tokens > baseline_estimate.total_tokens);
-    }
-
-    #[tokio::test]
-    async fn functional_prompt_returns_cancelled_when_token_is_pre_cancelled() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("should not be used"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(client.clone(), AgentConfig::default());
-        let token = CooperativeCancellationToken::new();
-        token.cancel();
-        agent.set_cancellation_token(Some(token));
-
-        let error = agent
-            .prompt("hello")
-            .await
-            .expect_err("prompt should cancel");
-        assert!(matches!(error, AgentError::Cancelled));
-        assert_eq!(client.requests.lock().await.len(), 0);
-    }
-
-    #[tokio::test]
-    async fn prompt_without_tools_completes_in_one_turn() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("Hello from model"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let new_messages = agent.prompt("hi").await.expect("prompt should succeed");
-
-        assert_eq!(new_messages.len(), 2);
-        assert_eq!(new_messages[0].role, MessageRole::User);
-        assert_eq!(new_messages[1].text_content(), "Hello from model");
-    }
-
-    #[tokio::test]
-    async fn functional_response_cache_reuses_model_response_for_identical_request() {
-        let calls = Arc::new(AtomicUsize::new(0));
-        let client = Arc::new(CountingStaticClient {
-            calls: calls.clone(),
-            response: ChatResponse {
-                message: Message::assistant_text("cached"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            },
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.append_message(Message::user("cache me"));
-        let baseline_messages = agent.messages.clone();
-        let start_index = baseline_messages.len().saturating_sub(1);
-
-        let _ = agent
-            .run_loop(start_index, None, false)
-            .await
-            .expect("first run should succeed");
-        agent.messages = baseline_messages;
-        let _ = agent
-            .run_loop(start_index, None, false)
-            .await
-            .expect("second run should succeed");
-
-        assert_eq!(calls.load(Ordering::Relaxed), 1);
-    }
-
-    #[tokio::test]
-    async fn regression_response_cache_disabled_dispatches_each_time() {
-        let calls = Arc::new(AtomicUsize::new(0));
-        let client = Arc::new(CountingStaticClient {
-            calls: calls.clone(),
-            response: ChatResponse {
-                message: Message::assistant_text("uncached"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            },
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                response_cache_enabled: false,
-                ..AgentConfig::default()
-            },
-        );
-        agent.append_message(Message::user("cache disabled"));
-        let baseline_messages = agent.messages.clone();
-        let start_index = baseline_messages.len().saturating_sub(1);
-
-        let _ = agent
-            .run_loop(start_index, None, false)
-            .await
-            .expect("first run should succeed");
-        agent.messages = baseline_messages;
-        let _ = agent
-            .run_loop(start_index, None, false)
-            .await
-            .expect("second run should succeed");
-
-        assert_eq!(calls.load(Ordering::Relaxed), 2);
-    }
-
-    #[tokio::test]
-    async fn regression_streaming_requests_bypass_response_cache() {
-        let calls = Arc::new(AtomicUsize::new(0));
-        let client = Arc::new(CountingStaticClient {
-            calls: calls.clone(),
-            response: ChatResponse {
-                message: Message::assistant_text("streamed"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            },
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.append_message(Message::user("streaming cache bypass"));
-        let baseline_messages = agent.messages.clone();
-        let start_index = baseline_messages.len().saturating_sub(1);
-        let sink = Arc::new(|_delta: String| {});
-
-        let _ = agent
-            .run_loop(start_index, Some(sink.clone()), false)
-            .await
-            .expect("first streamed run should succeed");
-        agent.messages = baseline_messages;
-        let _ = agent
-            .run_loop(start_index, Some(sink), false)
-            .await
-            .expect("second streamed run should succeed");
-
-        assert_eq!(calls.load(Ordering::Relaxed), 2);
-    }
-
-    #[tokio::test]
-    async fn prompt_executes_tool_calls_and_continues() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-
-        let second_assistant = Message::assistant_text("Done reading file");
-
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(ReadTool);
-
-        let new_messages = agent
-            .prompt("Read README.md")
-            .await
-            .expect("prompt should succeed");
-
-        assert_eq!(new_messages.len(), 4);
-        assert_eq!(new_messages[0].role, MessageRole::User);
-        assert_eq!(new_messages[1].role, MessageRole::Assistant);
-        assert_eq!(new_messages[2].role, MessageRole::Tool);
-        assert!(new_messages[2].text_content().contains("read:README.md"));
-        assert_eq!(new_messages[3].text_content(), "Done reading file");
-    }
-
-    #[tokio::test]
-    async fn integration_scoped_tool_lifecycle_supports_prompt_execution() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let second_assistant = Message::assistant_text("done");
-
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        assert!(!agent.has_tool("read"));
-
-        let messages = agent
-            .with_scoped_tool(ReadTool, |agent| {
-                Box::pin(async move { agent.prompt("read").await })
-            })
-            .await
-            .expect("scoped tool prompt should succeed");
-
-        assert!(
-            messages
-                .iter()
-                .any(|message| message.role == MessageRole::Tool),
-            "scoped tool should be available while running the closure"
-        );
-        assert!(!agent.has_tool("read"));
-    }
-
-    #[tokio::test]
-    async fn regression_scoped_tool_restores_replaced_tool_and_avoids_stale_cache() {
-        let make_tool_call = |id: &str| {
-            Message::assistant_blocks(vec![ContentBlock::ToolCall {
-                id: id.to_string(),
-                name: "cacheable_read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            }])
-        };
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: make_tool_call("call_1"),
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("base pass 1"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: make_tool_call("call_2"),
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("scoped pass"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: make_tool_call("call_3"),
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("base pass 2"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let base_calls = Arc::new(AtomicUsize::new(0));
-        let scoped_calls = Arc::new(AtomicUsize::new(0));
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(CacheableVariantTool {
-            label: "base",
-            calls: base_calls.clone(),
-        });
-
-        let first = agent
-            .prompt("first run")
-            .await
-            .expect("base tool run should succeed");
-        let first_tool = first
-            .iter()
-            .find(|message| message.role == MessageRole::Tool)
-            .expect("first tool result");
-        assert!(first_tool.text_content().contains("base:a.txt"));
-
-        let second = agent
-            .with_scoped_tool(
-                CacheableVariantTool {
-                    label: "scoped",
-                    calls: scoped_calls.clone(),
-                },
-                |agent| Box::pin(async move { agent.prompt("second run").await }),
-            )
-            .await
-            .expect("scoped tool run should succeed");
-        let second_tool = second
-            .iter()
-            .find(|message| message.role == MessageRole::Tool)
-            .expect("second tool result");
-        assert!(second_tool.text_content().contains("scoped:a.txt"));
-
-        let third = agent
-            .prompt("third run")
-            .await
-            .expect("restored base tool run should succeed");
-        let third_tool = third
-            .iter()
-            .find(|message| message.role == MessageRole::Tool)
-            .expect("third tool result");
-        assert!(third_tool.text_content().contains("base:a.txt"));
-
-        assert_eq!(base_calls.load(Ordering::Relaxed), 2);
-        assert_eq!(scoped_calls.load(Ordering::Relaxed), 1);
-    }
-
-    #[tokio::test]
-    async fn emits_expected_event_sequence_for_tool_turn() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let second_assistant = Message::assistant_text("done");
-
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(ReadTool);
-
-        let events = Arc::new(Mutex::new(Vec::<String>::new()));
-        let recorded = events.clone();
-        agent.subscribe(move |event| {
-            let label = match event {
-                AgentEvent::MessageAdded { message } => format!("message:{:?}", message.role),
-                AgentEvent::ToolExecutionStart { tool_name, .. } => {
-                    format!("tool_start:{tool_name}")
-                }
-                AgentEvent::ToolExecutionEnd { tool_name, .. } => format!("tool_end:{tool_name}"),
-                AgentEvent::TurnStart { turn } => format!("turn_start:{turn}"),
-                AgentEvent::TurnEnd { turn, .. } => format!("turn_end:{turn}"),
-                AgentEvent::ReplanTriggered { turn, .. } => format!("replan:{turn}"),
-                AgentEvent::CostUpdated { turn, .. } => format!("cost:{turn}"),
-                AgentEvent::CostBudgetAlert {
-                    threshold_percent, ..
-                } => format!("cost_alert:{threshold_percent}"),
-                AgentEvent::AgentStart => "agent_start".to_string(),
-                AgentEvent::AgentEnd { .. } => "agent_end".to_string(),
-            };
-
-            recorded
-                .lock()
-                .expect("event mutex should lock")
-                .push(label);
-        });
-
-        let _ = agent.prompt("read").await.expect("prompt should succeed");
-
-        let events = events.lock().expect("event mutex should lock").clone();
-        assert_eq!(
-            events,
-            vec![
-                "message:User",
-                "agent_start",
-                "turn_start:1",
-                "message:Assistant",
-                "tool_start:read",
-                "tool_end:read",
-                "message:Tool",
-                "turn_end:1",
-                "turn_start:2",
-                "message:Assistant",
-                "turn_end:2",
-                "agent_end",
-            ]
-        );
-    }
-
-    #[test]
-    fn unit_emit_isolates_panicking_handler_and_invokes_remaining_subscribers() {
-        let mut agent = Agent::new(Arc::new(EchoClient), AgentConfig::default());
-        let observed = Arc::new(AtomicUsize::new(0));
-
-        agent.subscribe(|event| {
-            if matches!(event, AgentEvent::AgentStart) {
-                panic!("forced event handler panic");
-            }
-        });
-        let observed_clone = observed.clone();
-        agent.subscribe(move |_event| {
-            observed_clone.fetch_add(1, Ordering::Relaxed);
-        });
-
-        agent.emit(AgentEvent::AgentStart);
-        assert_eq!(observed.load(Ordering::Relaxed), 1);
-    }
-
-    #[test]
-    fn unit_async_event_metrics_default_to_zero() {
-        let agent = Agent::new(Arc::new(EchoClient), AgentConfig::default());
-        assert_eq!(
-            agent.async_event_metrics(),
-            AsyncEventDispatchMetrics::default()
-        );
-    }
-
-    #[tokio::test]
-    async fn functional_async_subscriber_receives_events_and_records_metrics() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let observed = Arc::new(AtomicUsize::new(0));
-        let observed_clone = observed.clone();
-        agent.subscribe_async(move |_event| {
-            let observed_clone = observed_clone.clone();
-            async move {
-                observed_clone.fetch_add(1, Ordering::Relaxed);
-            }
-        });
-
-        let _ = agent.prompt("hello").await.expect("prompt should succeed");
-        let deadline = tokio::time::Instant::now() + Duration::from_millis(250);
-        while tokio::time::Instant::now() < deadline {
-            let metrics = agent.async_event_metrics();
-            if observed.load(Ordering::Relaxed) > 0 && metrics.completed > 0 {
-                break;
-            }
-            tokio::time::sleep(Duration::from_millis(5)).await;
-        }
-        assert!(
-            observed.load(Ordering::Relaxed) > 0,
-            "async handler should observe at least one event"
-        );
-        let metrics = agent.async_event_metrics();
-        assert!(metrics.enqueued > 0);
-        assert!(metrics.completed > 0);
-        assert_eq!(metrics.dropped_full, 0);
-    }
-
-    #[tokio::test]
-    async fn integration_async_subscriber_backpressure_drops_when_queue_is_full() {
-        let mut agent = Agent::new(
-            Arc::new(EchoClient),
-            AgentConfig {
-                async_event_queue_capacity: 1,
-                async_event_block_on_full: false,
-                async_event_handler_timeout_ms: None,
-                ..AgentConfig::default()
-            },
-        );
-        agent.subscribe_async(|_event| async move {
-            tokio::time::sleep(Duration::from_millis(80)).await;
-        });
-
-        for _ in 0..20 {
-            agent.emit(AgentEvent::AgentStart);
-        }
-        tokio::time::sleep(Duration::from_millis(250)).await;
-        let metrics = agent.async_event_metrics();
-        assert!(metrics.enqueued >= 1);
-        assert!(metrics.dropped_full > 0);
-    }
-
-    #[tokio::test]
-    async fn regression_async_subscriber_timeout_and_panic_are_isolated() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                async_event_queue_capacity: 16,
-                async_event_handler_timeout_ms: Some(20),
-                ..AgentConfig::default()
-            },
-        );
-        agent.subscribe_async(|_event| async move {
-            tokio::time::sleep(Duration::from_millis(100)).await;
-        });
-        agent.subscribe_async(|_event| async move {
-            panic!("forced async handler panic");
-        });
-
-        let _ = agent
-            .prompt("trigger async handlers")
-            .await
-            .expect("prompt should remain healthy");
-        tokio::time::sleep(Duration::from_millis(250)).await;
-        let metrics = agent.async_event_metrics();
-        assert!(metrics.timed_out > 0);
-        assert!(metrics.panicked > 0);
-    }
-
-    #[tokio::test]
-    async fn functional_prompt_completes_when_event_handler_panics() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.subscribe(|event| {
-            if matches!(event, AgentEvent::AgentStart) {
-                panic!("panic in functional handler");
-            }
-        });
-
-        let messages = agent
-            .prompt("hello")
-            .await
-            .expect("panic in event handler should not abort prompt");
-        assert_eq!(messages.last().expect("assistant").text_content(), "ok");
-    }
-
-    #[tokio::test]
-    async fn integration_tool_turn_completes_when_handler_panics_on_tool_events() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(ReadTool);
-        agent.subscribe(|event| {
-            if matches!(event, AgentEvent::ToolExecutionStart { .. }) {
-                panic!("panic on tool start");
-            }
-        });
-
-        let messages = agent
-            .prompt("read")
-            .await
-            .expect("tool turn should survive panicking handler");
-        assert!(
-            messages
-                .iter()
-                .any(|message| message.role == MessageRole::Tool),
-            "tool result should still be recorded"
-        );
-        assert_eq!(messages.last().expect("assistant").text_content(), "done");
-    }
-
-    #[tokio::test]
-    async fn regression_panicking_handler_does_not_break_subsequent_prompts() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: Message::assistant_text("first"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("second"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let event_count = Arc::new(AtomicUsize::new(0));
-        let event_count_clone = event_count.clone();
-        agent.subscribe(move |_event| {
-            event_count_clone.fetch_add(1, Ordering::Relaxed);
-        });
-        agent.subscribe(|event| {
-            if matches!(event, AgentEvent::AgentStart) {
-                panic!("panic every run");
-            }
-        });
-
-        let first = agent
-            .prompt("one")
-            .await
-            .expect("first prompt should succeed");
-        let second = agent
-            .prompt("two")
-            .await
-            .expect("second prompt should succeed");
-        assert_eq!(first.last().expect("assistant").text_content(), "first");
-        assert_eq!(second.last().expect("assistant").text_content(), "second");
-        assert!(
-            event_count.load(Ordering::Relaxed) > 0,
-            "non-panicking handler should keep receiving events across runs"
-        );
-    }
-
-    #[tokio::test]
-    async fn turn_end_events_include_usage_finish_reason_and_request_duration() {
-        let usage = ChatUsage {
-            input_tokens: 3,
-            output_tokens: 2,
-            total_tokens: 5,
-        };
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("done"),
-                finish_reason: Some("stop".to_string()),
-                usage: usage.clone(),
-            }])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let turn_ends = Arc::new(Mutex::new(Vec::<(
-            usize,
-            usize,
-            u64,
-            ChatUsage,
-            Option<String>,
-        )>::new()));
-        let captured = turn_ends.clone();
-        agent.subscribe(move |event| {
-            if let AgentEvent::TurnEnd {
-                turn,
-                tool_results,
-                request_duration_ms,
-                usage,
-                finish_reason,
-            } = event
-            {
-                captured.lock().expect("turn_end lock").push((
-                    *turn,
-                    *tool_results,
-                    *request_duration_ms,
-                    usage.clone(),
-                    finish_reason.clone(),
-                ));
-            }
-        });
-
-        let _ = agent.prompt("hello").await.expect("prompt should succeed");
-
-        let turn_ends = turn_ends.lock().expect("turn_end lock");
-        assert_eq!(turn_ends.len(), 1);
-        assert_eq!(turn_ends[0].0, 1);
-        assert_eq!(turn_ends[0].1, 0);
-        assert_eq!(turn_ends[0].3, usage);
-        assert_eq!(turn_ends[0].4.as_deref(), Some("stop"));
-    }
-
-    #[tokio::test]
-    async fn returns_max_turns_exceeded_for_infinite_tool_loop() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let second_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_2".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                max_turns: 2,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(ReadTool);
-
-        let error = agent.prompt("loop").await.expect_err("must hit max turns");
-        match error {
-            AgentError::MaxTurnsExceeded(2) => {}
-            other => panic!("expected AgentError::MaxTurnsExceeded(2), got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn rejects_invalid_tool_arguments_via_json_schema() {
-        let assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({}),
-        }]);
-
-        let final_assistant = Message::assistant_text("done");
-
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: final_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(ReadTool);
-
-        let messages = agent
-            .prompt("read without args")
-            .await
-            .expect("prompt succeeds");
-        let tool_message = messages
-            .iter()
-            .find(|message| message.role == MessageRole::Tool)
-            .expect("tool result must exist");
-        assert!(tool_message.is_error);
-        assert!(tool_message.text_content().contains("invalid arguments"));
-    }
-
-    #[tokio::test]
-    async fn integration_prompt_with_stream_emits_incremental_deltas() {
-        let client = Arc::new(StreamingMockClient {
-            response: ChatResponse {
-                message: Message::assistant_text("Hello"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            },
-            deltas: vec!["Hel".to_string(), "lo".to_string()],
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let streamed = Arc::new(Mutex::new(String::new()));
-        let sink_streamed = streamed.clone();
-        let sink = Arc::new(move |delta: String| {
-            sink_streamed
-                .lock()
-                .expect("stream lock")
-                .push_str(delta.as_str());
-        });
-
-        let new_messages = agent
-            .prompt_with_stream("hello", Some(sink))
-            .await
-            .expect("prompt should succeed");
-
-        assert_eq!(
-            new_messages
-                .last()
-                .expect("assistant message")
-                .text_content(),
-            "Hello"
-        );
-        assert_eq!(streamed.lock().expect("stream lock").as_str(), "Hello");
-    }
-
-    #[tokio::test]
-    async fn functional_streaming_retry_replays_buffer_without_duplicate_output() {
-        let attempts = Arc::new(AtomicUsize::new(0));
-        let client = Arc::new(RetryingStreamingClient {
-            outcomes: AsyncMutex::new(VecDeque::from([
-                RetryingStreamingOutcome {
-                    deltas: vec!["Hel".to_string()],
-                    response: Err(tau_ai::TauAiError::HttpStatus {
-                        status: 503,
-                        body: "transient".to_string(),
-                    }),
-                },
-                RetryingStreamingOutcome {
-                    deltas: vec!["Hello".to_string()],
-                    response: Ok(ChatResponse {
-                        message: Message::assistant_text("Hello"),
-                        finish_reason: Some("stop".to_string()),
-                        usage: ChatUsage::default(),
-                    }),
-                },
-            ])),
-            attempts: attempts.clone(),
-        });
-
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                request_max_retries: 1,
-                request_retry_initial_backoff_ms: 1,
-                request_retry_max_backoff_ms: 1,
-                stream_retry_with_buffering: true,
-                ..AgentConfig::default()
-            },
-        );
-        let streamed = Arc::new(Mutex::new(String::new()));
-        let streamed_sink = streamed.clone();
-        let sink = Arc::new(move |delta: String| {
-            streamed_sink
-                .lock()
-                .expect("stream lock")
-                .push_str(delta.as_str());
-        });
-
-        let messages = agent
-            .prompt_with_stream("hello", Some(sink))
-            .await
-            .expect("retrying stream should succeed");
-
-        assert_eq!(messages.last().expect("assistant").text_content(), "Hello");
-        assert_eq!(streamed.lock().expect("stream lock").as_str(), "Hello");
-        assert_eq!(attempts.load(Ordering::Relaxed), 2);
-    }
-
-    #[tokio::test]
-    async fn integration_streaming_retry_with_buffering_continues_tool_turns() {
-        let attempts = Arc::new(AtomicUsize::new(0));
-        let first_turn_retry = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let final_assistant = Message::assistant_text("done");
-        let client = Arc::new(RetryingStreamingClient {
-            outcomes: AsyncMutex::new(VecDeque::from([
-                RetryingStreamingOutcome {
-                    deltas: vec!["To".to_string()],
-                    response: Err(tau_ai::TauAiError::HttpStatus {
-                        status: 503,
-                        body: "temporary".to_string(),
-                    }),
-                },
-                RetryingStreamingOutcome {
-                    deltas: vec!["Tool ".to_string()],
-                    response: Ok(ChatResponse {
-                        message: first_turn_retry,
-                        finish_reason: Some("tool_calls".to_string()),
-                        usage: ChatUsage::default(),
-                    }),
-                },
-                RetryingStreamingOutcome {
-                    deltas: vec!["done".to_string()],
-                    response: Ok(ChatResponse {
-                        message: final_assistant,
-                        finish_reason: Some("stop".to_string()),
-                        usage: ChatUsage::default(),
-                    }),
-                },
-            ])),
-            attempts: attempts.clone(),
-        });
-
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                request_max_retries: 1,
-                request_retry_initial_backoff_ms: 1,
-                request_retry_max_backoff_ms: 1,
-                stream_retry_with_buffering: true,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(ReadTool);
-
-        let streamed = Arc::new(Mutex::new(String::new()));
-        let streamed_sink = streamed.clone();
-        let sink = Arc::new(move |delta: String| {
-            streamed_sink
-                .lock()
-                .expect("stream lock")
-                .push_str(delta.as_str());
-        });
-
-        let messages = agent
-            .prompt_with_stream("read file", Some(sink))
-            .await
-            .expect("streaming retry with tools should succeed");
-
-        assert_eq!(messages.last().expect("assistant").text_content(), "done");
-        assert!(
-            messages
-                .iter()
-                .any(|message| message.role == MessageRole::Tool),
-            "tool turn should still execute after a retried streaming failure"
-        );
-        assert_eq!(streamed.lock().expect("stream lock").as_str(), "Tool done");
-        assert_eq!(attempts.load(Ordering::Relaxed), 3);
-    }
-
-    #[tokio::test]
-    async fn regression_streaming_retry_disabled_fails_without_retrying_stream() {
-        let attempts = Arc::new(AtomicUsize::new(0));
-        let client = Arc::new(RetryingStreamingClient {
-            outcomes: AsyncMutex::new(VecDeque::from([
-                RetryingStreamingOutcome {
-                    deltas: vec!["Hel".to_string()],
-                    response: Err(tau_ai::TauAiError::HttpStatus {
-                        status: 503,
-                        body: "temporary".to_string(),
-                    }),
-                },
-                RetryingStreamingOutcome {
-                    deltas: vec!["Hello".to_string()],
-                    response: Ok(ChatResponse {
-                        message: Message::assistant_text("Hello"),
-                        finish_reason: Some("stop".to_string()),
-                        usage: ChatUsage::default(),
-                    }),
-                },
-            ])),
-            attempts: attempts.clone(),
-        });
-
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                request_max_retries: 1,
-                request_retry_initial_backoff_ms: 1,
-                request_retry_max_backoff_ms: 1,
-                stream_retry_with_buffering: false,
-                ..AgentConfig::default()
-            },
-        );
-        let streamed = Arc::new(Mutex::new(String::new()));
-        let streamed_sink = streamed.clone();
-        let sink = Arc::new(move |delta: String| {
-            streamed_sink
-                .lock()
-                .expect("stream lock")
-                .push_str(delta.as_str());
-        });
-
-        let error = agent
-            .prompt_with_stream("hello", Some(sink))
-            .await
-            .expect_err("disabled buffering should not retry streaming errors");
-        assert!(matches!(
-            error,
-            AgentError::Ai(tau_ai::TauAiError::HttpStatus { status: 503, .. })
-        ));
-        assert_eq!(streamed.lock().expect("stream lock").as_str(), "Hel");
-        assert_eq!(attempts.load(Ordering::Relaxed), 1);
-    }
-
-    #[tokio::test]
-    async fn functional_request_timeout_fails_closed_for_slow_provider() {
-        let mut agent = Agent::new(
-            Arc::new(DelayedEchoClient { delay_ms: 80 }),
-            AgentConfig {
-                request_max_retries: 0,
-                request_timeout_ms: Some(10),
-                ..AgentConfig::default()
-            },
-        );
-
-        let error = agent
-            .prompt("timeout please")
-            .await
-            .expect_err("slow provider should time out");
-        match error {
-            AgentError::RequestTimeout {
-                timeout_ms: 10,
-                attempt: 1,
-            } => {}
-            other => panic!("expected request timeout on first attempt, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn functional_token_budget_exceeded_fails_before_request_dispatch() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("should-not-run"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                system_prompt: String::new(),
-                max_estimated_input_tokens: Some(1),
-                ..AgentConfig::default()
-            },
-        );
-        let error = agent
-            .prompt("this prompt should exceed budget")
-            .await
-            .expect_err("token budget should fail closed");
-        assert!(matches!(error, AgentError::TokenBudgetExceeded { .. }));
-        assert!(
-            client.requests.lock().await.is_empty(),
-            "request should not be dispatched when budget check fails"
-        );
-    }
-
-    #[tokio::test]
-    async fn integration_total_token_budget_enforces_max_tokens_headroom() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("should-not-run"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                system_prompt: String::new(),
-                max_tokens: Some(64),
-                max_estimated_input_tokens: Some(10_000),
-                max_estimated_total_tokens: Some(30),
-                ..AgentConfig::default()
-            },
-        );
-        let error = agent
-            .prompt("small prompt")
-            .await
-            .expect_err("max_tokens should count against total budget");
-        match error {
-            AgentError::TokenBudgetExceeded {
-                max_total_tokens: 30,
-                ..
-            } => {}
-            other => panic!("expected total token budget failure, got {other:?}"),
-        }
-        assert!(
-            client.requests.lock().await.is_empty(),
-            "request should not be dispatched when total budget is exceeded"
-        );
-    }
-
-    #[test]
-    fn unit_estimate_usage_cost_usd_applies_input_and_output_rates() {
-        let usage = ChatUsage {
-            input_tokens: 2_000,
-            output_tokens: 500,
-            total_tokens: 2_500,
-        };
-        let cost = estimate_usage_cost_usd(&usage, Some(1.5), Some(6.0));
-        let expected = (2_000.0 * 1.5 + 500.0 * 6.0) / 1_000_000.0;
-        assert!((cost - expected).abs() < 1e-12);
-    }
-
-    #[test]
-    fn unit_normalize_cost_alert_thresholds_filters_invalid_and_deduplicates() {
-        assert_eq!(
-            normalize_cost_alert_thresholds(&[0, 80, 80, 120, 100]),
-            vec![80, 100]
-        );
-        assert_eq!(normalize_cost_alert_thresholds(&[]), vec![100]);
-    }
-
-    #[tokio::test]
-    async fn functional_prompt_emits_cost_update_event_when_model_pricing_present() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("done"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage {
-                    input_tokens: 200,
-                    output_tokens: 100,
-                    total_tokens: 300,
-                },
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                model_input_cost_per_million: Some(2.0),
-                model_output_cost_per_million: Some(4.0),
-                ..AgentConfig::default()
-            },
-        );
-        let observed = Arc::new(Mutex::new(Vec::<(usize, f64, f64, Option<f64>)>::new()));
-        let observed_clone = observed.clone();
-        agent.subscribe(move |event| {
-            if let AgentEvent::CostUpdated {
-                turn,
-                turn_cost_usd,
-                cumulative_cost_usd,
-                budget_usd,
-            } = event
-            {
-                observed_clone.lock().expect("events lock").push((
-                    *turn,
-                    *turn_cost_usd,
-                    *cumulative_cost_usd,
-                    *budget_usd,
-                ));
-            }
-        });
-
-        let _ = agent
-            .prompt("price this run")
-            .await
-            .expect("prompt should succeed");
-
-        let snapshot = agent.cost_snapshot();
-        let expected = (200.0 * 2.0 + 100.0 * 4.0) / 1_000_000.0;
-        assert_eq!(snapshot.input_tokens, 200);
-        assert_eq!(snapshot.output_tokens, 100);
-        assert_eq!(snapshot.total_tokens, 300);
-        assert!((snapshot.estimated_cost_usd - expected).abs() < 1e-12);
-        assert_eq!(snapshot.budget_usd, None);
-        assert_eq!(snapshot.budget_utilization, None);
-
-        let events = observed.lock().expect("events lock").clone();
-        assert_eq!(events.len(), 1);
-        assert_eq!(events[0].0, 1);
-        assert!((events[0].1 - expected).abs() < 1e-12);
-        assert!((events[0].2 - expected).abs() < 1e-12);
-        assert_eq!(events[0].3, None);
-    }
-
-    #[tokio::test]
-    async fn integration_budget_alerts_emit_once_per_threshold_across_multiple_prompts() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: Message::assistant_text("first"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage {
-                        input_tokens: 80_000,
-                        output_tokens: 0,
-                        total_tokens: 80_000,
-                    },
-                },
-                ChatResponse {
-                    message: Message::assistant_text("second"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage {
-                        input_tokens: 40_000,
-                        output_tokens: 0,
-                        total_tokens: 40_000,
-                    },
-                },
-                ChatResponse {
-                    message: Message::assistant_text("third"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage {
-                        input_tokens: 40_000,
-                        output_tokens: 0,
-                        total_tokens: 40_000,
-                    },
-                },
-            ])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                model_input_cost_per_million: Some(10.0),
-                model_output_cost_per_million: Some(0.0),
-                cost_budget_usd: Some(1.5),
-                cost_alert_thresholds_percent: vec![50, 80, 100],
-                ..AgentConfig::default()
-            },
-        );
-        let thresholds = Arc::new(Mutex::new(Vec::<u8>::new()));
-        let thresholds_clone = thresholds.clone();
-        agent.subscribe(move |event| {
-            if let AgentEvent::CostBudgetAlert {
-                threshold_percent, ..
-            } = event
-            {
-                thresholds_clone
-                    .lock()
-                    .expect("threshold lock")
-                    .push(*threshold_percent);
-            }
-        });
-
-        let _ = agent.prompt("step 1").await.expect("first prompt");
-        let _ = agent.prompt("step 2").await.expect("second prompt");
-        let _ = agent.prompt("step 3").await.expect("third prompt");
-
-        let snapshot = agent.cost_snapshot();
-        assert!((snapshot.estimated_cost_usd - 1.6).abs() < 1e-9);
-        assert_eq!(snapshot.budget_usd, Some(1.5));
-        let utilization = snapshot.budget_utilization.expect("utilization");
-        assert!(utilization > 1.0);
-
-        assert_eq!(
-            thresholds.lock().expect("threshold lock").as_slice(),
-            &[50, 80, 100]
-        );
-    }
-
-    #[tokio::test]
-    async fn regression_cost_budget_alert_threshold_normalization_avoids_duplicates() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("done"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage {
-                    input_tokens: 150_000,
-                    output_tokens: 0,
-                    total_tokens: 150_000,
-                },
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                model_input_cost_per_million: Some(10.0),
-                cost_budget_usd: Some(1.0),
-                cost_alert_thresholds_percent: vec![0, 80, 80, 120, 100],
-                ..AgentConfig::default()
-            },
-        );
-        let thresholds = Arc::new(Mutex::new(Vec::<u8>::new()));
-        let thresholds_clone = thresholds.clone();
-        agent.subscribe(move |event| {
-            if let AgentEvent::CostBudgetAlert {
-                threshold_percent, ..
-            } = event
-            {
-                thresholds_clone
-                    .lock()
-                    .expect("threshold lock")
-                    .push(*threshold_percent);
-            }
-        });
-
-        let _ = agent
-            .prompt("single run")
-            .await
-            .expect("prompt should succeed");
-        assert_eq!(
-            thresholds.lock().expect("threshold lock").as_slice(),
-            &[80, 100]
-        );
-    }
-
-    #[test]
-    fn unit_extract_json_payload_parses_plain_and_fenced_json() {
-        let plain = extract_json_payload(r#"{"ok":true,"count":2}"#).expect("plain json");
-        assert_eq!(plain["ok"], true);
-        assert_eq!(plain["count"], 2);
-
-        let fenced = extract_json_payload(
-            "result follows\n```json\n{\"status\":\"pass\",\"items\":[1,2]}\n```\nthanks",
-        )
-        .expect("fenced json");
-        assert_eq!(fenced["status"], "pass");
-        assert_eq!(fenced["items"][1], 2);
-    }
-
-    #[test]
-    fn unit_assistant_text_suggests_failure_matches_common_markers() {
-        assert!(assistant_text_suggests_failure(
-            "Unable to continue after the error."
-        ));
-        assert!(assistant_text_suggests_failure(
-            "I can't proceed with this tool."
-        ));
-        assert!(assistant_text_suggests_failure("   "));
-        assert!(!assistant_text_suggests_failure("Completed successfully."));
-    }
-
-    #[tokio::test]
-    async fn unit_vector_retrieval_prefers_semantically_related_entries() {
-        let history = vec![
-            Message::user("rust tokio async runtime troubleshooting"),
-            Message::assistant_text("pasta recipe with basil and tomato"),
-        ];
-        let matches = retrieve_memory_matches(
-            &history,
-            "tokio runtime async rust",
-            1,
-            64,
-            0.0,
-            &AgentConfig::default(),
-        )
-        .await;
-        assert_eq!(matches.len(), 1);
-        assert!(matches[0].text.contains("tokio"));
-
-        let query = embed_text_vector("tokio runtime async rust", 64);
-        let related = embed_text_vector("rust tokio async runtime troubleshooting", 64);
-        let unrelated = embed_text_vector("pasta recipe with basil and tomato", 64);
-        let related_score = query
-            .iter()
-            .zip(&related)
-            .map(|(left, right)| left * right)
-            .sum::<f32>();
-        let unrelated_score = query
-            .iter()
-            .zip(&unrelated)
-            .map(|(left, right)| left * right)
-            .sum::<f32>();
-        assert!(related_score > unrelated_score);
-    }
-
-    #[tokio::test]
-    async fn integration_memory_retrieval_uses_embedding_api_when_configured() {
-        let server = MockServer::start();
-        let embedding_mock = server.mock(|when, then| {
-            when.method(POST).path("/embeddings");
-            then.status(200)
-                .header("content-type", "application/json")
-                .body(
-                    r#"{
-  "data": [
-    { "embedding": [1.0, 0.0, 0.0] },
-    { "embedding": [0.95, 0.05, 0.0] },
-    { "embedding": [0.0, 1.0, 0.0] }
-  ]
-}"#,
-                );
-        });
-
-        let history = vec![
-            Message::assistant_text("retry strategy for postgres in payments service"),
-            Message::assistant_text("fresh tomato pasta recipe"),
-        ];
-        let config = AgentConfig {
-            memory_embedding_dimensions: 3,
-            memory_embedding_model: Some("text-embedding-3-small".to_string()),
-            memory_embedding_api_base: Some(server.url("")),
-            memory_embedding_api_key: Some("test-key".to_string()),
-            ..AgentConfig::default()
-        };
-
-        let matches = retrieve_memory_matches(
-            &history,
-            "payments postgres retry policy",
-            1,
-            3,
-            0.2,
-            &config,
-        )
-        .await;
-
-        embedding_mock.assert();
-        assert_eq!(matches.len(), 1);
-        assert!(matches[0].text.contains("payments"));
-        assert!(matches[0].text.contains("postgres"));
-    }
-
-    #[tokio::test]
-    async fn regression_memory_retrieval_falls_back_to_hash_when_embedding_api_fails() {
-        let server = MockServer::start();
-        let embedding_mock = server.mock(|when, then| {
-            when.method(POST).path("/embeddings");
-            then.status(500);
-        });
-
-        let history = vec![
-            Message::assistant_text("tokio runtime troubleshooting checklist"),
-            Message::assistant_text("basil and tomato pasta"),
-        ];
-        let config = AgentConfig {
-            memory_embedding_dimensions: 64,
-            memory_embedding_model: Some("text-embedding-3-small".to_string()),
-            memory_embedding_api_base: Some(server.url("")),
-            memory_embedding_api_key: Some("test-key".to_string()),
-            ..AgentConfig::default()
-        };
-
-        let matches = retrieve_memory_matches(&history, "tokio runtime", 1, 64, 0.0, &config).await;
-
-        embedding_mock.assert();
-        assert_eq!(matches.len(), 1);
-        assert!(matches[0].text.contains("tokio"));
-    }
-
-    #[test]
-    fn unit_build_structured_output_retry_prompt_includes_error_and_schema() {
-        let schema = serde_json::json!({
-            "type": "object",
-            "required": ["mode"]
-        });
-        let prompt =
-            build_structured_output_retry_prompt(&schema, "did not contain parseable JSON");
-        assert!(prompt.contains("did not contain parseable JSON"));
-        assert!(prompt.contains("\"required\":[\"mode\"]"));
-        assert!(prompt.contains("reply with only valid JSON"));
-    }
-
-    #[tokio::test]
-    async fn functional_prompt_json_returns_validated_value() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("{\"tasks\":[\"a\",\"b\"],\"ok\":true}"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "tasks": {
-                    "type": "array",
-                    "items": { "type": "string" }
-                },
-                "ok": { "type": "boolean" }
-            },
-            "required": ["tasks", "ok"],
-            "additionalProperties": false
-        });
-
-        let value = agent
-            .prompt_json("return tasks", &schema)
-            .await
-            .expect("structured output should succeed");
-        assert_eq!(value["ok"], true);
-        assert_eq!(value["tasks"][0], "a");
-    }
-
-    #[tokio::test]
-    async fn functional_prompt_json_retries_after_non_json_and_succeeds() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: Message::assistant_text("not-json-response"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("{\"tasks\":[\"retry\"],\"ok\":true}"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                structured_output_max_retries: 1,
-                ..AgentConfig::default()
-            },
-        );
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "tasks": {
-                    "type": "array",
-                    "items": { "type": "string" }
-                },
-                "ok": { "type": "boolean" }
-            },
-            "required": ["tasks", "ok"],
-            "additionalProperties": false
-        });
-
-        let value = agent
-            .prompt_json("return tasks", &schema)
-            .await
-            .expect("structured output retry should recover");
-        assert_eq!(value["ok"], true);
-        assert_eq!(value["tasks"][0], "retry");
-
-        let requests = client.requests.lock().await;
-        assert_eq!(requests.len(), 2, "prompt_json should perform one retry");
-        let retry_prompt = last_user_prompt(&requests[1]);
-        assert!(retry_prompt.contains("could not be accepted as structured JSON"));
-        assert!(retry_prompt.contains("\"tasks\""));
-    }
-
-    #[tokio::test]
-    async fn functional_prompt_json_enables_provider_json_mode_on_requests() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("{\"ok\":true}"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(client.clone(), AgentConfig::default());
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "ok": { "type": "boolean" }
-            },
-            "required": ["ok"],
-            "additionalProperties": false
-        });
-
-        let value = agent
-            .prompt_json("return ok", &schema)
-            .await
-            .expect("structured output should succeed");
-        assert_eq!(value["ok"], true);
-
-        let requests = client.requests.lock().await;
-        assert_eq!(requests.len(), 1);
-        assert!(requests[0].json_mode);
-    }
-
-    #[tokio::test]
-    async fn integration_prompt_json_accepts_fenced_json_payload() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text(
-                    "Here is the payload:\n```json\n{\"mode\":\"apply\",\"steps\":3}\n```",
-                ),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "mode": { "type": "string" },
-                "steps": { "type": "integer" }
-            },
-            "required": ["mode", "steps"]
-        });
-
-        let value = agent
-            .prompt_json("return mode", &schema)
-            .await
-            .expect("fenced structured output should parse");
-        assert_eq!(value["mode"], "apply");
-        assert_eq!(value["steps"], 3);
-    }
-
-    #[tokio::test]
-    async fn integration_continue_turn_json_retries_after_schema_failure_and_succeeds() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: Message::assistant_text("{\"mode\":\"apply\"}"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("{\"mode\":\"apply\",\"steps\":2}"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                structured_output_max_retries: 1,
-                ..AgentConfig::default()
-            },
-        );
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "mode": { "type": "string" },
-                "steps": { "type": "integer" }
-            },
-            "required": ["mode", "steps"]
-        });
-
-        let value = agent
-            .continue_turn_json(&schema)
-            .await
-            .expect("continue_turn_json should recover via retry");
-        assert_eq!(value["mode"], "apply");
-        assert_eq!(value["steps"], 2);
-
-        let requests = client.requests.lock().await;
-        assert_eq!(
-            requests.len(),
-            2,
-            "continue_turn_json should perform one retry request"
-        );
-        let retry_prompt = last_user_prompt(&requests[1]);
-        assert!(retry_prompt.contains("schema validation failed"));
-        assert!(retry_prompt.contains("\"steps\""));
-    }
-
-    #[tokio::test]
-    async fn integration_requests_with_registered_tools_use_auto_tool_choice() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("done"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(client.clone(), AgentConfig::default());
-        agent.register_tool(ReadTool);
-
-        let _ = agent.prompt("hello").await.expect("prompt should succeed");
-
-        let requests = client.requests.lock().await;
-        assert_eq!(requests.len(), 1);
-        assert_eq!(requests[0].tool_choice, Some(ToolChoice::Auto));
-        assert!(!requests[0].json_mode);
-    }
-
-    #[tokio::test]
-    async fn regression_prompt_json_fails_closed_on_non_json_response() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("not-json-response"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                structured_output_max_retries: 0,
-                ..AgentConfig::default()
-            },
-        );
-        let schema = serde_json::json!({ "type": "object" });
-
-        let error = agent
-            .prompt_json("return object", &schema)
-            .await
-            .expect_err("non-json output must fail");
-        assert!(matches!(error, AgentError::StructuredOutput(_)));
-        assert!(error.to_string().contains("did not contain parseable JSON"));
-    }
-
-    #[tokio::test]
-    async fn regression_prompt_json_fails_closed_on_schema_mismatch() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("{\"ok\":true}"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                structured_output_max_retries: 0,
-                ..AgentConfig::default()
-            },
-        );
-        let schema = serde_json::json!({
-            "type": "object",
-            "properties": {
-                "ok": { "type": "boolean" },
-                "tasks": { "type": "array" }
-            },
-            "required": ["ok", "tasks"]
-        });
-
-        let error = agent
-            .prompt_json("return object", &schema)
-            .await
-            .expect_err("schema mismatch must fail");
-        assert!(matches!(error, AgentError::StructuredOutput(_)));
-        assert!(error.to_string().contains("schema validation failed"));
-    }
-
-    #[tokio::test]
-    async fn regression_prompt_json_retry_exhaustion_fails_closed() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: Message::assistant_text("still-not-json"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: Message::assistant_text("again-not-json"),
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                structured_output_max_retries: 1,
-                ..AgentConfig::default()
-            },
-        );
-        let schema = serde_json::json!({ "type": "object" });
-
-        let error = agent
-            .prompt_json("return object", &schema)
-            .await
-            .expect_err("non-json output must fail after retries are exhausted");
-        assert!(matches!(error, AgentError::StructuredOutput(_)));
-        assert!(error.to_string().contains("did not contain parseable JSON"));
-
-        let requests = client.requests.lock().await;
-        assert_eq!(requests.len(), 2, "expected one retry attempt");
-    }
-
-    #[tokio::test]
-    async fn regression_requests_without_tools_keep_tool_choice_unset() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("done"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(client.clone(), AgentConfig::default());
-
-        let _ = agent.prompt("hello").await.expect("prompt should succeed");
-
-        let requests = client.requests.lock().await;
-        assert_eq!(requests.len(), 1);
-        assert_eq!(requests[0].tool_choice, None);
-    }
-
-    #[tokio::test]
-    async fn integration_tool_execution_cancellation_propagates_as_agent_cancelled() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "slow_read".to_string(),
-            arguments: serde_json::json!({
-                "path": "README.md"
-            }),
-        }]);
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: first_assistant,
-                finish_reason: Some("tool_calls".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(SlowReadTool { delay_ms: 500 });
-        let token = CooperativeCancellationToken::new();
-        agent.set_cancellation_token(Some(token.clone()));
-
-        let cancel_task = tokio::spawn(async move {
-            tokio::time::sleep(Duration::from_millis(50)).await;
-            token.cancel();
-        });
-
-        let error = agent
-            .prompt("read with cancellation")
-            .await
-            .expect_err("prompt should cancel cooperatively");
-        assert!(
-            matches!(error, AgentError::Cancelled),
-            "expected AgentError::Cancelled, got {error:?}"
-        );
-        cancel_task.await.expect("cancel task should complete");
-    }
-
-    #[tokio::test]
-    async fn regression_agent_can_continue_after_cancellation_token_is_cleared() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(client, AgentConfig::default());
-        let token = CooperativeCancellationToken::new();
-        token.cancel();
-        agent.set_cancellation_token(Some(token));
-
-        let error = agent
-            .prompt("cancelled run")
-            .await
-            .expect_err("first prompt should be cancelled");
-        assert!(matches!(error, AgentError::Cancelled));
-
-        agent.set_cancellation_token(None);
-        let new_messages = agent
-            .prompt("second run")
-            .await
-            .expect("agent should continue once token is cleared");
-        assert_eq!(
-            new_messages
-                .last()
-                .expect("assistant response should exist")
-                .text_content(),
-            "ok"
-        );
-    }
-
-    #[tokio::test]
-    async fn regression_continue_turn_json_fails_closed_when_assistant_lacks_json() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ack"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                structured_output_max_retries: 0,
-                ..AgentConfig::default()
-            },
-        );
-        let schema = serde_json::json!({ "type": "object" });
-
-        let error = agent
-            .continue_turn_json(&schema)
-            .await
-            .expect_err("missing json must fail");
-        assert!(matches!(error, AgentError::StructuredOutput(_)));
-    }
-
-    #[tokio::test]
-    async fn functional_replan_prompt_injected_after_failed_tool_and_failure_response() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({}),
-        }]);
-        let second_assistant =
-            Message::assistant_text("I cannot continue because the tool failed.");
-        let third_assistant = Message::assistant_text("recovered");
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: third_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                react_max_replans_on_tool_failure: 1,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(ReadTool);
-        let replan_count = Arc::new(AtomicUsize::new(0));
-        let replan_count_sink = replan_count.clone();
-        agent.subscribe(move |event| {
-            if matches!(event, AgentEvent::ReplanTriggered { .. }) {
-                replan_count_sink.fetch_add(1, Ordering::Relaxed);
-            }
-        });
-
-        let messages = agent
-            .prompt("read")
-            .await
-            .expect("replan flow should recover");
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "recovered"
-        );
-        assert_eq!(replan_count.load(Ordering::Relaxed), 1);
-
-        let requests = client.requests.lock().await;
-        assert_eq!(
-            requests.len(),
-            3,
-            "expected replan to trigger an extra turn"
-        );
-        let replan_prompt = last_user_prompt(&requests[2]);
-        assert!(replan_prompt.contains("One or more tool calls failed"));
-    }
-
-    #[tokio::test]
-    async fn functional_request_messages_attach_memory_recall_for_relevant_history() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                max_context_messages: Some(4),
-                memory_retrieval_limit: 2,
-                memory_embedding_dimensions: 64,
-                memory_min_similarity: 0.2,
-                ..AgentConfig::default()
-            },
-        );
-        agent.append_message(Message::user(
-            "postgres retry configuration for orders service",
-        ));
-        agent.append_message(Message::assistant_text(
-            "increase postgres pool size for orders workloads",
-        ));
-        agent.append_message(Message::user("cache ttl cleanup"));
-        agent.append_message(Message::assistant_text("set ttl to 15m"));
-
-        let _ = agent
-            .prompt("postgres orders service retry policy")
-            .await
-            .expect("prompt should succeed");
-
-        let requests = client.requests.lock().await;
-        let first_request = requests.first().expect("request should be captured");
-        let recall = first_request
-            .messages
-            .iter()
-            .find(|message| {
-                message.role == MessageRole::System
-                    && message.text_content().starts_with(MEMORY_RECALL_PREFIX)
-            })
-            .expect("memory recall system message should be attached");
-        assert!(recall.text_content().contains("postgres"));
-        assert!(recall.text_content().contains("orders"));
-    }
-
-    #[tokio::test]
-    async fn integration_parallel_tool_execution_runs_calls_concurrently_and_preserves_order() {
-        let first_assistant = Message::assistant_blocks(vec![
-            ContentBlock::ToolCall {
-                id: "call_1".to_string(),
-                name: "slow_read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            },
-            ContentBlock::ToolCall {
-                id: "call_2".to_string(),
-                name: "slow_read".to_string(),
-                arguments: serde_json::json!({ "path": "b.txt" }),
-            },
-        ]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                max_parallel_tool_calls: 2,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(SlowReadTool { delay_ms: 120 });
-
-        let started = Instant::now();
-        let messages = agent
-            .prompt("read both")
-            .await
-            .expect("prompt should succeed");
-        let elapsed = started.elapsed();
-
-        assert!(
-            elapsed < Duration::from_millis(230),
-            "expected concurrent tool execution under 230ms, got {elapsed:?}"
-        );
-
-        let tool_messages = messages
-            .iter()
-            .filter(|message| message.role == MessageRole::Tool)
-            .collect::<Vec<_>>();
-        assert_eq!(tool_messages.len(), 2);
-        assert!(tool_messages[0].text_content().contains("read:a.txt"));
-        assert!(tool_messages[1].text_content().contains("read:b.txt"));
-    }
-
-    #[tokio::test]
-    async fn integration_parallel_tool_cache_reuses_results_across_chunks() {
-        let first_assistant = Message::assistant_blocks(vec![
-            ContentBlock::ToolCall {
-                id: "call_1".to_string(),
-                name: "counting_read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            },
-            ContentBlock::ToolCall {
-                id: "call_2".to_string(),
-                name: "counting_read".to_string(),
-                arguments: serde_json::json!({ "path": "b.txt" }),
-            },
-            ContentBlock::ToolCall {
-                id: "call_3".to_string(),
-                name: "counting_read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            },
-        ]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-        let calls = Arc::new(AtomicUsize::new(0));
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                max_parallel_tool_calls: 2,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(CountingReadTool {
-            calls: calls.clone(),
-            cacheable: true,
-        });
-
-        let messages = agent
-            .prompt("read all")
-            .await
-            .expect("prompt should succeed");
-        let tool_messages = messages
-            .iter()
-            .filter(|message| message.role == MessageRole::Tool)
-            .collect::<Vec<_>>();
-        assert_eq!(tool_messages.len(), 3);
-        assert!(tool_messages[0].text_content().contains("counting:a.txt"));
-        assert!(tool_messages[1].text_content().contains("counting:b.txt"));
-        assert!(tool_messages[2].text_content().contains("counting:a.txt"));
-        assert_eq!(calls.load(Ordering::Relaxed), 2);
-    }
-
-    #[tokio::test]
-    async fn regression_non_cacheable_tool_executes_each_identical_call() {
-        let first_assistant = Message::assistant_blocks(vec![
-            ContentBlock::ToolCall {
-                id: "call_1".to_string(),
-                name: "counting_read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            },
-            ContentBlock::ToolCall {
-                id: "call_2".to_string(),
-                name: "counting_read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            },
-        ]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-        let calls = Arc::new(AtomicUsize::new(0));
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                max_parallel_tool_calls: 1,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(CountingReadTool {
-            calls: calls.clone(),
-            cacheable: false,
-        });
-
-        let _ = agent
-            .prompt("read duplicate")
-            .await
-            .expect("prompt should succeed");
-        assert_eq!(calls.load(Ordering::Relaxed), 2);
-    }
-
-    #[tokio::test]
-    async fn integration_replan_flow_can_recover_with_follow_up_tool_call() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({}),
-        }]);
-        let second_assistant = Message::assistant_text("Unable to continue after that tool error.");
-        let third_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_2".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let fourth_assistant = Message::assistant_text("done after replan");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: third_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: fourth_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                react_max_replans_on_tool_failure: 1,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(ReadTool);
-
-        let messages = agent
-            .prompt("read")
-            .await
-            .expect("replan flow should recover with second tool call");
-        let tool_messages = messages
-            .iter()
-            .filter(|message| message.role == MessageRole::Tool)
-            .collect::<Vec<_>>();
-        assert_eq!(tool_messages.len(), 2);
-        assert!(tool_messages[0].is_error);
-        assert!(!tool_messages[1].is_error);
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "done after replan"
-        );
-    }
-
-    #[tokio::test]
-    async fn integration_memory_recall_ranks_relevant_entries_ahead_of_unrelated_entries() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                max_context_messages: Some(2),
-                memory_retrieval_limit: 1,
-                memory_embedding_dimensions: 64,
-                memory_min_similarity: 0.1,
-                ..AgentConfig::default()
-            },
-        );
-        agent.append_message(Message::user("rust tokio runtime diagnostics"));
-        agent.append_message(Message::user("pasta recipe tomato basil"));
-        agent.append_message(Message::assistant_text("acknowledged"));
-
-        let _ = agent
-            .prompt("tokio runtime troubleshooting")
-            .await
-            .expect("prompt should succeed");
-
-        let requests = client.requests.lock().await;
-        let first_request = requests.first().expect("request should be captured");
-        let recall = first_request
-            .messages
-            .iter()
-            .find(|message| {
-                message.role == MessageRole::System
-                    && message.text_content().starts_with(MEMORY_RECALL_PREFIX)
-            })
-            .expect("memory recall message should exist");
-        assert!(recall.text_content().contains("tokio"));
-        assert!(!recall.text_content().contains("pasta recipe"));
-    }
-
-    #[tokio::test]
-    async fn regression_bug_1_max_parallel_tool_calls_zero_clamps_to_safe_serial_execution() {
-        let first_assistant = Message::assistant_blocks(vec![
-            ContentBlock::ToolCall {
-                id: "call_1".to_string(),
-                name: "read".to_string(),
-                arguments: serde_json::json!({ "path": "a.txt" }),
-            },
-            ContentBlock::ToolCall {
-                id: "call_2".to_string(),
-                name: "read".to_string(),
-                arguments: serde_json::json!({ "path": "b.txt" }),
-            },
-        ]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                max_parallel_tool_calls: 0,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(ReadTool);
-
-        let messages = agent
-            .prompt("read both")
-            .await
-            .expect("zero parallel limit should be normalized to a safe value");
-        let tool_messages = messages
-            .iter()
-            .filter(|message| message.role == MessageRole::Tool)
-            .collect::<Vec<_>>();
-        assert_eq!(tool_messages.len(), 2);
-        assert!(tool_messages[0].text_content().contains("read:a.txt"));
-        assert!(tool_messages[1].text_content().contains("read:b.txt"));
-    }
-
-    #[tokio::test]
-    async fn regression_no_replan_when_assistant_reports_success_after_tool_failure() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({}),
-        }]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                react_max_replans_on_tool_failure: 1,
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(ReadTool);
-        let replan_count = Arc::new(AtomicUsize::new(0));
-        let replan_count_sink = replan_count.clone();
-        agent.subscribe(move |event| {
-            if matches!(event, AgentEvent::ReplanTriggered { .. }) {
-                replan_count_sink.fetch_add(1, Ordering::Relaxed);
-            }
-        });
-
-        let messages = agent
-            .prompt("read")
-            .await
-            .expect("prompt should complete without forced replan");
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "done"
-        );
-        assert_eq!(replan_count.load(Ordering::Relaxed), 0);
-        assert_eq!(client.requests.lock().await.len(), 2);
-    }
-
-    #[tokio::test]
-    async fn regression_memory_recall_disabled_when_limit_is_zero() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                max_context_messages: Some(2),
-                memory_retrieval_limit: 0,
-                ..AgentConfig::default()
-            },
-        );
-        agent.append_message(Message::user("postgres connection issue"));
-        agent.append_message(Message::assistant_text("ack"));
-        agent.append_message(Message::user("retry strategy"));
-
-        let _ = agent.prompt("postgres retry policy").await.expect("prompt");
-        let requests = client.requests.lock().await;
-        let first_request = requests.first().expect("request should be captured");
-        assert!(first_request
-            .messages
-            .iter()
-            .all(|message| !message.text_content().starts_with(MEMORY_RECALL_PREFIX)));
-    }
-
-    #[tokio::test]
-    async fn functional_context_window_limits_request_messages_and_compacts_history() {
-        let client = Arc::new(CapturingMockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-            requests: AsyncMutex::new(Vec::new()),
-        });
-
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                max_context_messages: Some(4),
-                ..AgentConfig::default()
-            },
-        );
-        agent.append_message(Message::user("u1"));
-        agent.append_message(Message::assistant_text("a1"));
-        agent.append_message(Message::user("u2"));
-        agent.append_message(Message::assistant_text("a2"));
-        agent.append_message(Message::user("u3"));
-
-        let _ = agent.prompt("latest").await.expect("prompt should succeed");
-
-        let requests = client.requests.lock().await;
-        let first_request = requests.first().expect("request should be captured");
-        assert_eq!(first_request.messages.len(), 4);
-        assert_eq!(first_request.messages[0].role, MessageRole::System);
-        assert_eq!(first_request.messages[1].role, MessageRole::System);
-        assert!(first_request.messages[1]
-            .text_content()
-            .starts_with(CONTEXT_SUMMARY_PREFIX));
-        assert_eq!(first_request.messages[2].text_content(), "u3");
-        assert_eq!(first_request.messages[3].text_content(), "latest");
-        assert!(
-            agent.messages().len() <= 4,
-            "history should be compacted to configured context window"
-        );
-    }
-
-    #[test]
-    fn unit_bounded_messages_inserts_summary_with_system_prompt() {
-        let messages = vec![
-            Message::system("sys"),
-            Message::user("u1"),
-            Message::assistant_text("a1"),
-            Message::user("u2"),
-            Message::assistant_text("a2"),
-        ];
-
-        let bounded = bounded_messages(&messages, 4);
-        assert_eq!(bounded.len(), 4);
-        assert_eq!(bounded[0].role, MessageRole::System);
-        assert_eq!(bounded[1].role, MessageRole::System);
-        assert!(bounded[1]
-            .text_content()
-            .starts_with(CONTEXT_SUMMARY_PREFIX));
-        assert_eq!(bounded[2].text_content(), "u2");
-        assert_eq!(bounded[3].text_content(), "a2");
-    }
-
-    #[test]
-    fn regression_bounded_messages_inserts_summary_without_system_prompt() {
-        let messages = vec![
-            Message::user("u1"),
-            Message::assistant_text("a1"),
-            Message::user("u2"),
-            Message::assistant_text("a2"),
-        ];
-
-        let bounded = bounded_messages(&messages, 3);
-        assert_eq!(bounded.len(), 3);
-        assert_eq!(bounded[0].role, MessageRole::System);
-        assert!(bounded[0]
-            .text_content()
-            .starts_with(CONTEXT_SUMMARY_PREFIX));
-        assert_eq!(bounded[1].text_content(), "u2");
-        assert_eq!(bounded[2].text_content(), "a2");
-    }
-
-    #[test]
-    fn regression_truncate_chars_preserves_utf8_and_appends_ellipsis() {
-        let long = "alpha beta gamma delta epsilon zeta eta theta";
-        let truncated = truncate_chars(long, 12);
-        assert_eq!(truncated.chars().count(), 12);
-        assert!(truncated.ends_with('…'));
-
-        let long_unicode = "hello 👋 from τau runtime";
-        let truncated_unicode = truncate_chars(long_unicode, 9);
-        assert_eq!(truncated_unicode.chars().count(), 9);
-        assert!(truncated_unicode.ends_with('…'));
-
-        let very_long = "x".repeat(CONTEXT_SUMMARY_MAX_CHARS + 200);
-        let clipped = truncate_chars(&very_long, CONTEXT_SUMMARY_MAX_CHARS);
-        assert!(clipped.chars().count() <= CONTEXT_SUMMARY_MAX_CHARS);
-    }
-
-    #[tokio::test]
-    async fn regression_retry_transient_request_failures_and_recover_response() {
-        let client = Arc::new(RetryThenSuccessClient {
-            remaining_failures: AsyncMutex::new(1),
-            attempts: AsyncMutex::new(0),
-            response: ChatResponse {
-                message: Message::assistant_text("recovered"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            },
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                request_max_retries: 2,
-                request_retry_initial_backoff_ms: 1,
-                request_retry_max_backoff_ms: 2,
-                ..AgentConfig::default()
-            },
-        );
-
-        let messages = agent
-            .prompt("retry please")
-            .await
-            .expect("prompt should recover");
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "recovered"
-        );
-        assert_eq!(*client.attempts.lock().await, 2);
-    }
-
-    #[tokio::test]
-    async fn regression_request_timeout_retries_and_recovers_when_next_attempt_is_fast() {
-        let client = Arc::new(TimeoutThenSuccessClient {
-            delays_ms: AsyncMutex::new(VecDeque::from([40, 0])),
-            attempts: AsyncMutex::new(0),
-            response: ChatResponse {
-                message: Message::assistant_text("timeout-recovered"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            },
-        });
-        let mut agent = Agent::new(
-            client.clone(),
-            AgentConfig {
-                request_max_retries: 1,
-                request_retry_initial_backoff_ms: 1,
-                request_retry_max_backoff_ms: 1,
-                request_timeout_ms: Some(10),
-                ..AgentConfig::default()
-            },
-        );
-
-        let messages = agent
-            .prompt("recover after timeout")
-            .await
-            .expect("second attempt should succeed");
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "timeout-recovered"
-        );
-        assert_eq!(*client.attempts.lock().await, 2);
-    }
-
-    #[tokio::test]
-    async fn regression_token_budget_none_disables_estimation_gate() {
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([ChatResponse {
-                message: Message::assistant_text("ok"),
-                finish_reason: Some("stop".to_string()),
-                usage: ChatUsage::default(),
-            }])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                system_prompt: String::new(),
-                max_estimated_input_tokens: None,
-                max_estimated_total_tokens: None,
-                ..AgentConfig::default()
-            },
-        );
-        let oversized_prompt = "x".repeat(250_000);
-        let messages = agent
-            .prompt(oversized_prompt)
-            .await
-            .expect("token gate disabled should allow prompt");
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "ok"
-        );
-    }
-
-    #[tokio::test]
-    async fn regression_tool_panic_isolated_to_error_tool_result() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "panic_tool".to_string(),
-            arguments: serde_json::json!({}),
-        }]);
-        let second_assistant = Message::assistant_text("continued");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut agent = Agent::new(client, AgentConfig::default());
-        agent.register_tool(PanicTool);
-
-        let messages = agent.prompt("panic").await.expect("prompt should continue");
-        let tool_message = messages
-            .iter()
-            .find(|message| message.role == MessageRole::Tool)
-            .expect("tool result should be present");
-        assert!(tool_message.is_error);
-        assert!(tool_message
-            .text_content()
-            .contains("execution task failed"));
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "continued"
-        );
-    }
-
-    #[tokio::test]
-    async fn integration_tool_timeout_returns_error_tool_message_and_continues_turn() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "slow_read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let second_assistant = Message::assistant_text("continued after timeout");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-        let mut agent = Agent::new(
-            client,
-            AgentConfig {
-                tool_timeout_ms: Some(10),
-                ..AgentConfig::default()
-            },
-        );
-        agent.register_tool(SlowReadTool { delay_ms: 75 });
-
-        let messages = agent
-            .prompt("slow read")
-            .await
-            .expect("prompt should continue after tool timeout");
-        let tool_message = messages
-            .iter()
-            .find(|message| message.role == MessageRole::Tool)
-            .expect("tool result should be present");
-        assert!(tool_message.is_error);
-        assert!(tool_message.text_content().contains("timed out after 10ms"));
-        assert_eq!(
-            messages.last().expect("assistant response").text_content(),
-            "continued after timeout"
-        );
-    }
-
-    #[tokio::test]
-    async fn unit_agent_fork_clones_state_without_aliasing_messages() {
-        let first_assistant = Message::assistant_blocks(vec![ContentBlock::ToolCall {
-            id: "call_1".to_string(),
-            name: "read".to_string(),
-            arguments: serde_json::json!({ "path": "README.md" }),
-        }]);
-        let second_assistant = Message::assistant_text("done");
-        let client = Arc::new(MockClient {
-            responses: AsyncMutex::new(VecDeque::from([
-                ChatResponse {
-                    message: first_assistant,
-                    finish_reason: Some("tool_calls".to_string()),
-                    usage: ChatUsage::default(),
-                },
-                ChatResponse {
-                    message: second_assistant,
-                    finish_reason: Some("stop".to_string()),
-                    usage: ChatUsage::default(),
-                },
-            ])),
-        });
-
-        let mut base = Agent::new(client, AgentConfig::default());
-        base.register_tool(ReadTool);
-        base.append_message(Message::user("seed message"));
-
-        let mut fork = base.fork();
-        let fork_messages = fork.prompt("read").await.expect("fork prompt");
-        assert!(
-            fork_messages
-                .iter()
-                .any(|message| message.role == MessageRole::Tool),
-            "fork should inherit registered tools and execute tool calls"
-        );
-        assert_eq!(base.messages().len(), 2);
-        assert_eq!(fork.messages().len(), 6);
-    }
-
-    #[tokio::test]
-    async fn integration_run_parallel_prompts_executes_runs_concurrently_with_ordered_results() {
-        let agent = Agent::new(
-            Arc::new(DelayedEchoClient { delay_ms: 90 }),
-            AgentConfig::default(),
-        );
-
-        let started = Instant::now();
-        let results = agent
-            .run_parallel_prompts(vec!["prompt-1", "prompt-2", "prompt-3", "prompt-4"], 4)
-            .await;
-        let elapsed = started.elapsed();
-
-        assert!(
-            elapsed < Duration::from_millis(260),
-            "expected concurrent runs under 260ms, got {elapsed:?}"
-        );
-        assert_eq!(results.len(), 4);
-
-        for (index, result) in results.into_iter().enumerate() {
-            let messages = result.expect("parallel run should succeed");
-            assert_eq!(messages[0].role, MessageRole::User);
-            assert_eq!(
-                messages.last().expect("assistant reply").text_content(),
-                format!("echo:prompt-{}", index + 1)
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn integration_bug_6_run_parallel_prompts_allows_zero_parallel_limit() {
-        let agent = Agent::new(Arc::new(EchoClient), AgentConfig::default());
-        let results = agent.run_parallel_prompts(vec!["p1", "p2", "p3"], 0).await;
-        assert_eq!(results.len(), 3);
-        for (index, result) in results.into_iter().enumerate() {
-            let messages = result.expect("zero parallel limit should clamp to a valid value");
-            assert_eq!(
-                messages.last().expect("assistant reply").text_content(),
-                format!("echo:p{}", index + 1)
-            );
-        }
-    }
-
-    #[tokio::test]
-    async fn regression_run_parallel_prompts_isolates_failures_per_prompt() {
-        let agent = Agent::new(
-            Arc::new(SelectiveFailureEchoClient),
-            AgentConfig {
-                request_max_retries: 0,
-                ..AgentConfig::default()
-            },
-        );
-
-        let results = agent
-            .run_parallel_prompts(vec!["ok-1", "fail-2", "ok-3"], 2)
-            .await;
-
-        assert_eq!(results.len(), 3);
-        assert!(results[0].as_ref().is_ok());
-        assert!(matches!(
-            results[1],
-            Err(AgentError::Ai(tau_ai::TauAiError::HttpStatus {
-                status: 503,
-                ..
-            }))
-        ));
-        assert!(results[2].as_ref().is_ok());
-    }
-
-    #[tokio::test]
-    async fn functional_run_parallel_prompts_returns_empty_for_empty_input() {
-        let agent = Agent::new(Arc::new(EchoClient), AgentConfig::default());
-        let results = agent
-            .run_parallel_prompts(std::iter::empty::<&str>(), 4)
-            .await;
-        assert!(results.is_empty());
-    }
+    #[path = "structured_output_and_parallel.rs"]
+    mod structured_output_and_parallel;
 }
